refactored common functionality out into mixin.
