+++ app/components/course-objective-list.js
@@ -1,48 +1,14 @@
 import Ember from 'ember';
-import SortableByPosition from 'ilios/mixins/sortable-by-position';
+import SortableObjectiveList from 'ilios/mixins/sortable-objective-list';
 
-const { computed, Component, RSVP } = Ember;
-const { Promise, all } = RSVP;
+const { Component, computed } = Ember;
+const { alias } = computed;
 
-export default Component.extend(SortableByPosition, {
+export default Component.extend(SortableObjectiveList, {
   classNames: ['course-objective-list'],
   objectivesForRemovalConfirmation: [],
   editable: true,
-  isSorting: false,
-  isSaving: false,
-  course: null,
-  totalObjectivesToSave: null,
-  currentObjectivesSaved: null,
-
-  objectives: computed('course.objectives.[]', function(){
-    return this.get('course').get('objectives');
-  }),
-
-  hasMoreThanOneObjective: computed('objectives.[]', function(){
-    return new Promise(resolve => {
-      this.get('objectives').then(objectives => {
-        resolve(objectives.length > 1);
-      });
-    });
-  }),
-
-  sortedObjectives: computed('objectives.[]', function(){
-    return new Promise(resolve => {
-      this.get('objectives').then(objectives => {
-        resolve(objectives.toArray().sort(this.get('positionSortingCallback')));
-      });
-    });
-  }),
-
-  saveSomeObjectives(arr){
-    let chunk = arr.splice(0, 5);
-    return all(chunk.invoke('save')).then(() => {
-      if (arr.length){
-        this.set('currentObjectivesSaved', this.get('currentObjectivesSaved') + chunk.length);
-        return this.saveSomeObjectives(arr);
-      }
-    });
-  },
+  course: alias('subject'),
 
   actions: {
     remove(objective){
@@ -55,24 +21,5 @@ export default Component.extend(SortableByPosition, {
     confirmRemoval(objective){
       this.get('objectivesForRemovalConfirmation').pushObject(objective.get('id'));
     },
-    saveSortOrder(objectives){
-      this.set('isSaving', true);
-      for (let i = 0, n = objectives.length; i < n; i++) {
-        let lm = objectives[i];
-        lm.set('position', i + 1);
-      }
-      this.set('totalObjectivesToSave', objectives.length);
-      this.set('currentObjectivesSaved', 0);
-
-      this.saveSomeObjectives(objectives).then(() => {
-        this.set('isSaving', false);
-        this.set('isSorting', false);
-      });
-    },
-
-    cancelSorting() {
-      this.set('isSorting', false);
-    },
-
   }
 });
