+++ app/components/programyear-objective-list.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 
 const { computed, Component } = Ember;
-const { sort, not } = computed;
+const { not } = computed;
 
 export default Component.extend({
   classNames: ['programyear-objective-list'],
@@ -10,8 +10,4 @@ export default Component.extend({
   objectives: computed('programYear.objectives.[]', function(){
     return this.get('programYear').get('objectives');
   }),
-  sortedObjectives: sort('objectives', function(a, b){
-    return parseInt(a.get( 'id' )) - parseInt(b.get( 'id' ));
-  }),
-
 });
