+++ app/templates/components/detail-objectives.hbs
@@ -41,7 +41,7 @@
   {{#unless isManaging class="horizontal"}}
     {{#if isCourse}}
       {{course-objective-list
-        course=subject
+        subject=subject
         editable=editable
         manageParents=(action "manageParents")
         manageDescriptors=(action "manageDescriptors")
